% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance_rate.R
\name{acceptance_rate}
\alias{acceptance_rate}
\title{Function for getting the acceptance rate by institution}
\usage{
acceptance_rate(institution_id)
}
\arguments{
\item{institution_id}{Institution id. Find matching name and ids in the \code{dir_info2020} dataset.}
}
\value{
A statement with the acceptance rate for inputted institution.
}
\description{
Provides the acceptance rate for the inputted institution
}
\examples{
acceptance_rate(167835)
acceptance_rate(100830)
}
