% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adm2020}
\alias{adm2020}
\title{Admissions and Test Scores 2020 Data}
\format{
A data frame of XXX rows representing hourly measurements and 40 variables

\describe{
\item{UNITID}{Unique identification number of the institution}
\item{ADMCON1}{Secondary school GPA}
}
}
\source{
Download from Integrated Postsecondary Education Data System,
\url{https://nces.ed.gov/ipeds/datacenter/data/ADM2020.zip}.
}
\usage{
adm2020
}
\description{
This file contains information about the undergraduate selection process for entering first-time, degree/certificate-seeking students. This includes information about admission considerations,  applicants, applicants that were admitted, and admitted students who enrolled. SAT and ACT test scores are included for institutions, that require test scores for admission. These data are applicable for institutions that do not have an open admissions policy for entering first-time students. Beginning with 2016, writing scores for both SAT and ACT are no longer collected. The possible values for the admission consideration variables have changed from data prior to 2016.  ""Do not know"" is no longer an option and ""considered but not required"" was added.  See frequencies for complete listing.
}
\keyword{datasets}
