% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin2020}
\alias{admin2020}
\title{Administration 2020 Data}
\format{
A data frame of  63,843 rows representing full-time instructional staff on the payroll of the institution and 34 variables

\describe{
\item{INSTITUTION_ID}{Unique identification number of the institution}
\item{STAFF_CAT}{Instructional staff category (1	= All full-time instructional staff; 100 = Instructional staff with faculty status; 101 = Instructional staff, professors; 102 = Instructional staff, associate professors; 103	= Instructional staff, assistant professors; 104 = Instructional staff, instructors; 105 = Instructional staff, lecturers; 106 = Instructional staff, no academic rank; 200 = Tenured total; 201 = Tenured, professors; 202	= Tenured, associate professors; 203 = Tenured, assistant professors; 204 = Tenured, instructors; 205 = Tenured, lecturers; 206	= Tenured, no academic rank; 300 = On-Tenure track total; 301	= On-tenure track, professors; 302 = On-tenure track, associate professors; 303	= On-tenure track, assistant professors; 304 = On-tenure track, instructors; 305 = On-tenure track, lecturers; 306 = On-tenure track, no academic rank; 400	= Not on tenure track/No tenure system system total; 401 = Not on tenure/no tenure system, professors; 402 = Not on tenure/no tenure system, associate professors; 403 = Not on tenure/no tenure system, assistant professors; 404 = Not on tenure/no tenure system, instructors; 405	= Not on tenure/no tenure system, lecturers; 406 = Not on tenure/no tenure system, no academic rank; 410 = Not on tenure track/no tenure system, multi-year and indefinite contracts, total; 411 = Not on tenure track/no tenure system, multi-year and indefinite contracts, professors; 412	= Not on tenure track/no tenure system, multi-year and indefinite contracts, asssociate professors; 413	= Not on tenure track/no tenure system, multi-year and indefinite contracts, assistant professors; 414 = Not on tenure track/no tenure system, multi-year and indefinite contracts, instructors; 415 = Not on tenure track/no tenure system, multi-year and indefinite contracts contract,lecturers; 416 = Not on tenure track/no tenure system, multi-year and indefinite contracts, no academic rank; 420	= Not on tenure track/no tenure system, annual contract, total; 421	= Not on tenure track/no tenure system, annual contract, professors; 422 = Not on tenure track/no tenure system, annual contract, asssociate professors; 423 = Not on tenure track/no tenure system, annual contract, assistant professors; 424	= Not on tenure track/no tenure system, annual contract, instructors; 425	= Not on tenure track/no tenure system, annual contract,lecturers; 426 = Not on tenure track/no tenure system, annual contract, no academic rank; 430	= Not on tenure track/no tenure system, less-than-annual contract, total; 431	= Not on tenure track/no tenure system, less-than-annual contract, professors; 432 = Not on tenure track/no tenure system, less-than-annual contract, asssociate professors; 433 = Not on tenure track/no tenure system, less-than-annual contract, assistant professors; 434	= Not on tenure track/no tenure system, less-than-annual contract, instructors; 435	= Not on tenure track/no tenure system, less-than-annual contract,lecturers; 436 = Not on tenure track/no tenure system, less-than-annual contract, no academic rank; 440	= Not on tenure track/no tenure system, multi-year contract, total; 441	= Not on tenure track/no tenure system, multi-year contract, professors; 442 = Not on tenure track/no tenure system, multi-year contract, associate professors; 443	= Not on tenure track/no tenure system, multi-year contract, assistant professors; 444 = Not on tenure track/no tenure system, multi-year contract, instructors; 445 = Not on tenure track/no tenure system, multi-year contract, lecturers; 446 = Not on tenure track/no tenure system, multi-year contract, no academic rank; 450	= Not on tenure track/no tenure system, multi-year contract, total; 451	= Not on tenure track/no tenure system, indefinite contract, professors; 452 = Not on tenure track/no tenure system, indefinite contract, associate professors; 453	= Not on tenure track/no tenure system, indefinite contract, assistant professors; 454 = Not on tenure track/no tenure system, indefinite contract, instructors; 455 = Not on tenure track/no tenure system, indefinite contract, lecturers; 456 = Not on tenure track/no tenure system, indefinite contract, no academic rank; 500	= Without faculty status)}
\item{FACSTAT}{Faculty and tenure status (0 = All full-time instructional staff; 10 = With faculty status, total; 20	= With faculty status, tenured; 30 = With faculty status, on tenure track; 40	= With faculty status not on tenure track/No tenure system, total; 41	= With faculty status not on tenure track/No tenure system, multi-year and indefinite contract; 44 = With faculty status not on tenure track/No tenure system, multi-year contract; 45 = With faculty status not on tenure track/No tenure system, indefinite contract; 42 = With faculty status not on tenure track/No tenure system, annual contract; 43 = With faculty status not on tenure track/No tenure system, less-than-annual contract; 50 = Without faculty status)}
\item{ARANK: a factor with 7 levels: \verb{All ranks}, \code{Professors}, \verb{Associate professors}, \verb{Assistant professors}, \code{Instructors}, \code{Lecturers} and \verb{No academic rank}}{Academic rank}
\item{TOTAL}{Grand total individuals}
\item{TOTAL_M}{Grand total men}
\item{TOTAL_W}{Grand total women}
\item{NATIVE_TOTAL}{American Indian or Alaska Native total}
\item{NATIVE_MEN}{American Indian or Alaska Native men}
\item{NATIVE_WOMEN}{American Indian or Alaska Native women}
\item{ASIAN_TOTAL}{Asian total}
\item{ASIAN_M}{Asian men}
\item{ASIAN_W}{Asian men}
\item{BLACK_TOTAL}{Black or African American total}
\item{BLACK_M}{Black or African American men}
\item{BLACK_W}{Black or African American women}
\item{HISP_TOTAL}{Hispanic or Latino total}
\item{HISP_M}{Hispanic or Latino men}
\item{HISP_W}{Hispanic or Latino women}
\item{NHPI_TOTAL}{Native Hawaiian or Other Pacific Islander total}
\item{NHPI_M}{Native Hawaiian or Other Pacific Islander men}
\item{NHPI_W}{Native Hawaiian or Other Pacific Islander women}
\item{WHITE_TOTAL}{White total}
\item{WHITE_M}{White men}
\item{WHITE_W}{White women}
\item{MULT_TOTAL}{Two or more races total}
\item{MULT_M}{Two or more races men}
\item{MULT_W}{Two or more races women}
\item{UNKNOWN_TOTAL}{Race/ethnicity unknown total}
\item{UNKNOWN_M}{Race/ethnicity unknown men}
\item{UNKNOWN_W}{Race/ethnicity unknown women}
\item{NRA_TOTAL}{Nonresident alien total}
\item{NRA_M}{Nonresident alien men}
\item{NRA_W}{Nonresident alien women}
}
}
\source{
Download from Integrated Postsecondary Education Data System,
\url{https://nces.ed.gov/ipeds/datacenter/data/S2020_IS.zip}.
}
\usage{
admin2020
}
\description{
This file contains the number of full-time instructional staff on the payroll of the institution as of November 1,  by faculty and tenure status, academic rank, race/ethnicity and gender. This file has multiple records per institution.  Each record is uniquely defined by the variables IPEDS ID (INSTIUTION_ID). These data are applicable to degree-granting institutions with 15 or more full-time employees and related administrative offices.
}
\details{
Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions.
}
\keyword{datasets}
