% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admission_reqs.R
\name{admission_reqs}
\alias{admission_reqs}
\title{Function for Admission Requirements for an Institution}
\usage{
admission_reqs(institution_id)
}
\arguments{
\item{institution_id}{Institution ID. Find these on the \code{dir_info2020} dataset.}
}
\value{
A tibble of the requirements for the inputted institution.
}
\description{
Lists and orders requirements for admission by priority
}
\examples{
admission_reqs(167835)
admission_reqs(130794)
}
