% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{cipcodes}
\alias{cipcodes}
\title{Classification of Instructional Programs (CIP) Codes}
\format{
a data.frame with 2,318 observations of 8 variables.
}
\description{
The Classification of Instructional Programs (CIP) provides a taxonomic 
scheme that supports the accurate tracking and reporting of fields of study 
and program completions activity. CIP was originally developed by the 
U.S. Department of Education's National Center for Education Statistics (NCES) 
in 1980, with revisions occurring in 1985, 1990, and 2000.
}
\references{
National Center for Educational Statistics: https://nces.ed.gov/ipeds/cipcode
}
\keyword{data}
