% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_int.R
\name{compare_int}
\alias{compare_int}
\title{Function for Comparing (2) Institutions Side by Side, Chosen by User Input}
\usage{
compare_int(Institution1_ID, Institution2_ID)
}
\arguments{
\item{Institution1_ID}{(numeric) 1st Institution chosen by user to be compared against Institution2 (100000-500000)}

\item{Institution2_ID}{(numeric) 2nd Institution chosen by user to be compared against Institution1 (100000-500000)}
}
\value{
A 2 column table with the qualities of the 2 institutions being compared in the rows.
}
\description{
Outputs Table Comparing Various Qualities Each Institution Has
}
\examples{
compare_int(Institution1_ID = 100654, Institution2_ID = 100663)
compare_int(Institution1_ID = "University of California Los Angeles", Institution2_ID = "University of California Berkeley")
}
