% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conference}
\alias{conference}
\title{Conferences Data}
\format{
A data frame of 500 rows representing conferences and 3 variables

\describe{
\item{Numeric_Code}{Numeric code representing a conference}
\item{Conference}{Explanation of numeric code}
\item{Sport}{Type of sport for each conference}
}
}
\usage{
conference
}
\description{
This file contains the details for conferences in the Offerings 2020 (\code{offerings2020}) dataset.
}
\keyword{datasets}
