% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk of 2000 CIP codes to 2010 CIP codes
 
Maps the 2000 CIP codes with the 2010 CIP codes.}
\format{a data.frame with 2,202 observations of 6 variables.}
\description{
Crosswalk of 2000 CIP codes to 2010 CIP codes
 
Maps the 2000 CIP codes with the 2010 CIP codes.
}
\references{
NCES https://nces.ed.gov/ipeds/cipcode/crosswalk.aspx?y=55
}
\keyword{data}
