% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_IPEDS.R
\name{download_ipeds}
\alias{download_ipeds}
\title{Downloads IPEDS data for a given year.}
\usage{
download_ipeds(year = as.integer(format(Sys.Date(), "\%Y")) - 1,
  dir = getOption("ipeds.download.dir"), useProvisional = TRUE,
  force = FALSE, cleanup = FALSE, ...)
}
\arguments{
\item{year}{the year of the survey}

\item{dir}{the directory to download the data to. Defaults to the package directory.}

\item{useProvisional}{if TRUE, provisional data files will be downloaded if 
the final version is not available.}

\item{force}{if TRUE, the function will redownload the file.}

\item{cleanup}{if TRUE, the zip and MS Access files will be deleted after the Rda file is created.}

\item{...}{other parameters passed to url_exists.}
}
\value{
TRUE if the data has been successfully downloaded.
}
\description{
This function will download IPEDS data in MS Access file format. By default,
files are downloaded to a subfolder of the ipeds package in the system's
R library directory. Once downloaded, the MS Access database will be converted
to a R data file for faster loading in subsequent R sessions.
}
