\name{downloadAllSurveys}
\alias{downloadAllSurveys}
\title{downloadAllSurveys}
\usage{
  downloadAllSurveys(year, ...)
}
\arguments{
  \item{year}{the year to download}

  \item{...}{other parameters passed to
  \code{\link{downloadIPEDSSurvey}}}
}
\description{
  Downloads all available surveys for the given year.
}

