% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipedsHelp.R
\name{downloadHelp}
\alias{downloadHelp}
\title{Download the data dictionary for the given year.}
\usage{
downloadHelp(year)
}
\arguments{
\item{year}{the IPEDS year to download the data dictionary.}
}
\description{
Internal function.
}
