% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIPEDS.R
\name{downloadIPEDS}
\alias{downloadIPEDS}
\title{Downloads IPEDS data for a given year.}
\usage{
downloadIPEDS(year = as.integer(format(Sys.Date(), "\%Y")) - 1,
  useProvisional = TRUE, dir = paste0(system.file(package = "ipeds"),
  "/data/downloaded/"), force = FALSE, ...)
}
\arguments{
\item{year}{the year of the survey}

\item{dir}{the directory to download the data to. Defaults to the package directory.}

\item{force}{if TRUE, the function will redownload the file.}

\item{...}{other parmaters}
}
\value{
TRUE if the data has been successfully downloaded.
}
\description{
Downloads IPEDS data for a given year.
}
