% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallEnrollment.R
\name{fallEnrollment}
\alias{fallEnrollment}
\title{Internal method.}
\usage{
fallEnrollment(data, mappings = NULL, institutionId = NA,
  institutionName = NA, documentId = "a",
  createdDate = format(Sys.time(), "\%Y-\%m-\%dT\%H:\%M:\%S"),
  noteMessage = NA, transmissionType = "Original",
  documentType = "IPEDS")
}
\description{
Internal method.
}
