% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fin_aid1920}
\alias{fin_aid1920}
\title{Student Financial Aid Data 2019 - 2020}
\format{
A data frame of 5859 rows representing institutions and 91 variables

\describe{
\item{INSTITUTION_ID}{Unique identification number of the institution}
\item{UG_NUM}{Total number of undergraduates - full-year cohort}
\item{PCT_AID}{Percent of undergraduate students awarded federal, state, local, institutional or other sources of grant aid}
\item{AVG_AID}{Average amount of federal, state, local, institutional or other sources of grant aid awarded to undergraduate students}
\item{AVG_PELL}{Average amount Pell grant aid awarded to undergraduate students}
\item{AVG_FLOAN}{Average amount of federal student loans awarded to undergraduate students}
\item{AVG_AWRD}{Average amount of federal, state, local or institutional grant aid awarded}
\item{AVG_FGRANT_FTFT}{Average amount of federal grant aid awarded to full-time first-time undergraduates}
\item{AVG_PGRANT_FTFT}{Average amount of Pell grant aid awarded to full-time first-time undergraduates}
\item{AVG_FOTHER_FTFT}{Average amount of other federal grant aid awarded to full-time first-time undergraduates}
\item{AVG_STGRANT_FTFT}{Average amount of state/local grant aid awarded to full-time first-time undergraduates}
\item{AVG_INST_AID_FTFT}{Average amount of institutional grant aid awarded to full-time first-time undergraduates}
\item{AVG_LOAN_FTFT}{Average amount of student loans awarded to full-time first-time undergraduates}
\item{AVG_FLOAN_FTFT}{Average amount of federal student loans awarded to full-time first-time undergraduates}
\item{AVG_OLOAN_FTFT}{Average amount of other student loans awarded to full-time first-time undergraduates}
\item{IS_AVG_GAID_1920}{Average amount of grant and scholarship aid awarded, 2019-20}
\item{IS_AVG_GAID_1819}{Average amount of grant and scholarship aid awarded, 2018-19}
\item{IS_AVG_GAID_1718}{Average amount of grant and scholarship aid awarded, 2017-18}
\item{AVG_AID_AL_1920}{Average amount of grant and scholarship aid awarded, all income levels, 2019-20}
\item{AVG_AID_L1_1920}{Average amount of grant and scholarship aid awarded, income level (0-30,000), 2019-20}
\item{AVG_AID_L2_1920}{Average amount of grant and scholarship aid awarded, income level (30,001-48,000), 2019-20}
\item{AVG_AID_L3_1920}{Average amount of grant and scholarship aid awarded, income level (48,001-75,000), 2019-20}
\item{AVG_AID_L4_1920}{Average amount of grant and scholarship aid awarded, income level (75,001-110,000), 2019-20}
\item{AVG_AID_L5_1920}{Average amount of grant and scholarship aid awarded, income level (110,001 or more), 2019-20}
\item{AVG_AID_AL_1819}{Average amount of grant and scholarship aid awarded, all income levels, 2018-19}
\item{AVG_AID_L1_1819}{Average amount of grant and scholarship aid awarded, income level (0-30,000), 2018-19}
\item{AVG_AID_L2_1819}{Average amount of grant and scholarship aid awarded, income level (30,001-48,000), 2018-19}
\item{AVG_AID_L3_1819}{Average amount of grant and scholarship aid awarded, income level (48,001-75,000), 2018-19}
\item{AVG_AID_L4_1819}{Average amount of grant and scholarship aid awarded, income level (75,001-110,000), 2018-19}
\item{AVG_AID_L5_1819}{Average amount of grant and scholarship aid awarded, income level (110,001 or more), 2018-19}
\item{AVG_AID_AL_1718}{Average amount of grant and scholarship aid awarded, all income levels, 2017-18}
\item{AVG_AID_L1_1718}{Average amount of grant and scholarship aid awarded, income level (0-30,000), 2017-18}
\item{AVG_AID_L2_1718}{Average amount of grant and scholarship aid awarded, income level (30,001-48,000), 2017-18}
\item{AVG_AID_L3_1718}{Average amount of grant and scholarship aid awarded, income level (48,001-75,000), 2017-18}
\item{AVG_AID_L4_1718}{Average amount of grant and scholarship aid awarded, income level (75,001-110,000), 2017-18}
\item{AVG_AID_L5_1718}{Average amount of grant and scholarship aid awarded, income level (110,001 or more), 2017-18}
\item{AVG_NPRICE_1920}{Average net price-students awarded grant or scholarship aid, 2019-20}
\item{AVG_NPRICE_1819}{Average net price-students awarded grant or scholarship aid, 2018-19}
\item{AVG_NPRICE_1718}{Average net price-students awarded grant or scholarship aid, 2017-18}
\item{AVG_NPRICE_L1_1920}{Average net price (income 0-30,000)-students awarded Title IV federal financial aid, 2019-20}
\item{AVG_NPRICE_L2_1920}{Average net price (income 30,001-48,000)-students awarded Title IV federal financial aid, 2019-20}
\item{AVG_NPRICE_L3_1920}{Average net price (income 48,001-75,000)-students awarded Title IV federal financial aid, 2019-20}
\item{AVG_NPRICE_L4_1920}{Average net price (income 75,001-110,000)-students awarded Title IV federal financial aid, 2019-20}
\item{AVG_NPRICE_L5_1920}{Average net price (income over 110,000)-students awarded Title IV federal financial aid, 2019-20}
\item{AVG_NPRICE_L1_1819}{Average net price (income 0-30,000)-students awarded Title IV federal financial aid, 2018-19}
\item{AVG_NPRICE_L2_1819}{Average net price (income 30,001-48,000)-students awarded Title IV federal financial aid, 2018-19}
\item{AVG_NPRICE_L3_1819}{Average net price (income 48,001-75,000)-students awarded Title IV federal financial aid, 2018-19}
\item{AVG_NPRICE_L4_1819}{Average net price (income 75,001-110,000)-students awarded Title IV federal financial aid, 2018-19}
\item{AVG_NPRICE_L5_1819}{Average net price (income over 110,000)-students awarded Title IV federal financial aid, 2018-19}
\item{AVG_NPRICE_L1_1718}{Average net price (income 0-30,000)-students awarded Title IV federal financial aid, 2017-18}
\item{AVG_NPRICE_L2_1718}{Average net price (income 30,001-48,000)-students awarded Title IV federal financial aid, 2017-18}
\item{AVG_NPRICE_L3_1718}{Average net price (income 48,001-75,000)-students awarded Title IV federal financial aid, 2017-18}
\item{AVG_NPRICE_L4_1718}{Average net price (income 75,001-110,000)-students awarded Title IV federal financial aid, 2017-18}
\item{AVG_NPRICE_L5_1718}{Average net price (income over 110,000)-students awarded Title IV federal financial aid, 2017-18}
\item{OS_AVG_GAID_1718}{Average amount of grant and scholarship aid awarded, 2017-18}
\item{OS_AVG_GAID_AL_1920}{Average amount of grant and scholarship aid awarded, all income levels, 2019-20}
\item{OS_AVG_GAID_L1_1920}{Average amount of grant and scholarship aid awarded, income level (0-30,000), 2019-20}
\item{OS_AVG_GAID_L2_1920}{Average amount of grant and scholarship aid awarded, income level (30,001-48,000), 2019-20}
\item{OS_AVG_GAID_L3_1920}{Average amount of grant and scholarship aid awarded, income level (48,001-75,000), 2019-20}
\item{OS_AVG_GAID_L4_1920}{Average amount of grant and scholarship aid awarded, income level (75,001-110,000), 2019-20}
\item{OS_AVG_GAID_L5_1920}{Average amount of grant and scholarship aid awarded, income level (110,001 or more), 2019-20}
\item{OS_AVG_GAID_AL_1819}{Average amount of grant and scholarship aid awarded, all income levels, 2018-19}
\item{OS_AVG_GAID_L1_1819}{Average amount of grant and scholarship aid awarded, income level (0-30,000), 2018-19}
\item{OS_AVG_GAID_L2_1819}{Average amount of grant and scholarship aid awarded, income level (30,001-48,000), 2018-19}
\item{OS_AVG_GAID_L3_1819}{Average amount of grant and scholarship aid awarded, income level (48,001-75,000), 2018-19}
\item{OS_AVG_GAID_L4_1819}{Average amount of grant and scholarship aid awarded, income level (75,001-110,000), 2018-19}
\item{OS_AVG_GAID_L5_1819}{Average amount of grant and scholarship aid awarded, income level (110,001 or more), 2018-19}
\item{OS_AVG_GAID_AL_1718}{Average amount of grant and scholarship aid awarded, all income levels, 2017-18}
\item{OS_AVG_GAID_L1_1718}{Average amount of grant and scholarship aid awarded, income level (0-30,000), 2017-18}
\item{OS_AVG_GAID_L2_1718}{Average amount of grant and scholarship aid awarded, income level (30,001-48,000), 2017-18}
\item{OS_AVG_GAID_L3_1718}{Average amount of grant and scholarship aid awarded, income level (48,001-75,000), 2017-18}
\item{OS_AVG_GAID_L4_1718}{Average amount of grant and scholarship aid awarded, income level (75,001-110,000), 2017-18}
\item{OS_AVG_GAID_L5_1718}{Average amount of grant and scholarship aid awarded, income level (110,001 or more), 2017-18}
\item{OS_AVG_NPRICE_1920}{Average net price-students awarded grant or scholarship aid, 2019-20}
\item{OS_AVG_NPRICE_1819}{Average net price-students awarded grant or scholarship aid, 2018-19}
\item{OS_AVG_NPRICE_1718}{Average net price-students awarded grant or scholarship aid, 2017-18}
\item{OS_AVG_NPRICE_L1_1920}{Average net price (income 0-30,000)-students awarded Title IV federal financial aid, 2019-20}
\item{OS_AVG_NPRICE_L2_1920}{Average net price (income 30,001-48,000)-students awarded Title IV federal financial aid, 2019-20}
\item{OS_AVG_NPRICE_L3_1920}{Average net price (income 48,001-75,000)-students awarded Title IV federal financial aid, 2019-20}
\item{OS_AVG_NPRICE_L4_1920}{Average net price (income 75,001-110,000)-students awarded Title IV federal financial aid, 2019-20}
\item{OS_AVG_NPRICE_L5_1920}{Average net price (income over 110,000)-students awarded Title IV federal financial aid, 2019-20}
\item{OS_AVG_NPRICE_L1_1819}{Average net price (income 0-30,000)-students awarded Title IV federal financial aid, 2018-19}
\item{OS_AVG_NPRICE_L2_1819}{Average net price (income 30,001-48,000)-students awarded Title IV federal financial aid, 2018-19}
\item{OS_AVG_NPRICE_L3_1819}{Average net price (income 48,001-75,000)-students awarded Title IV federal financial aid, 2018-19}
\item{OS_AVG_NPRICE_L4_1819}{Average net price (income 75,001-110,000)-students awarded Title IV federal financial aid, 2018-19}
\item{OS_AVG_NPRICE_L5_1819}{Average net price (income over 110,000)-students awarded Title IV federal financial aid, 2018-19}
\item{OS_AVG_NPRICE_L1_1718}{Average net price (income 0-30,000)-students awarded Title IV federal financial aid, 2017-18}
\item{OS_AVG_NPRICE_L2_1718}{Average net price (income 30,001-48,000)-students awarded Title IV federal financial aid, 2017-18}
\item{OS_AVG_NPRICE_L3_1718}{Average net price (income 48,001-75,000)-students awarded Title IV federal financial aid, 2017-18}
\item{OS_AVG_NPRICE_L4_1718}{Average net price (income 75,001-110,000)-students awarded Title IV federal financial aid, 2017-18}
\item{OS_AVG_NPRICE_L5_1718}{Average net price (income over 110,000)-students awarded Title IV federal financial aid, 2017-18}
}
}
\source{
Download from Integrated Postsecondary Education Data System,
\url{https://nces.ed.gov/ipeds/datacenter/data/SFA1920.zip}.
}
\usage{
fin_aid1920
}
\description{
Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions.
}
\keyword{datasets}
