% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYearString.R
\name{formatYear}
\alias{formatYear}
\title{Formats the year for a specific survey.}
\usage{
formatYear(surveyId, year)
}
\arguments{
\item{surveyId}{the survey ID from the data(surveys) data.frame.}

\item{year}{the IPEDS release year.}
}
\description{
Returns the formatted year for the given survey.
}
