% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{getIPEDSDownloadDirectory}
\alias{getIPEDSDownloadDirectory}
\title{Private function that returns the location to download IPEDS data files.}
\usage{
getIPEDSDownloadDirectory()
}
\value{
directory to download IPEDS data.
}
\description{
This will check to see if the ipeds.download.dir option has been set, if it
has that will be returned, if not it will set the option to the current
the data/downloaded/ directory of the current working directory, i.e.:
paste0(system.file(package="ipeds"), '/data/downloaded/')
}
