\name{getIPEDSSurvey}
\alias{getIPEDSSurvey}
\title{getIPEDSSurvey}
\description{
  This function will return the IPEDS survey data for the given survey and year. Note that the first time this function is invoked it will attempt to download the data file from the IPEDS website.
}
\usage{
}
\arguments{
  \item{surveyId}{The survey ID of the dictionary to display (i.e. surveys$SurveyID)}
  \item{year}{The year of the dictionary to display.}
}
\details{
}
\value{
}
\references{
  http://nces.ed.gov/ipeds/datacenter/
}
\author{
  Jason Bryer <jbryer@excelsior.edu>
}
\note{
}


\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ completions }

