% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYearString.R
\name{getYearString}
\alias{getYearString}
\title{Converts the year to the IPEDS academic year format.}
\usage{
getYearString(year)
}
\arguments{
\item{year}{the IPEDS release year.}
}
\value{
a string with the cohort year
}
\description{
IPEDS uses a YYYY-YY format for dates in files. For 2018 data release, the
date format would be 2017-18.
}
