% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_help.R
\name{ipeds_help}
\alias{ipeds_help}
\title{Returns data.frame with the data dictionary.}
\usage{
ipeds_help(table, year = as.integer(format(Sys.Date(), "\%Y")) - 1,
  dir = getIPEDSDownloadDirectory())
}
\arguments{
\item{table}{if specified, the variable information for the given survey result table is returned.}

\item{year}{the year of the survey}

\item{dir}{the directory to download the data to. Defaults to the package directory.}
}
\description{
This will return a data.frame with a list of all the tables available for the given
IEPDS year. Or, if the table parameter is specified, a data.frame with all the
variables available in that table.
}
