% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_survey.R
\name{ipeds_survey}
\alias{ipeds_survey}
\title{Returns}
\usage{
ipeds_survey(year = as.integer(format(Sys.Date(), "\%Y")) - 1,
  dir = paste0(system.file(package = "ipeds"), "/data/downloaded/"),
  table)
}
\arguments{
\item{year}{the year of the survey}

\item{dir}{the directory to download the data to. Defaults to the package directory.}

\item{table}{the IPEDS survey result to return.}
}
\description{
Returns
}
