% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ipeds.R
\name{load_ipeds}
\alias{load_ipeds}
\title{Loads IPEDS data for a given year.}
\usage{
load_ipeds(year = as.integer(format(Sys.Date(), "\%Y")) - 1,
  dir = paste0(system.file(package = "ipeds"), "/data/downloaded/"))
}
\arguments{
\item{year}{the year of the survey}

\item{dir}{the directory to download the data to. Defaults to the package directory.}
}
\value{
a list with all tables (as data.frames) for the given year.
}
\description{
This function will return a large list with all the tables for a given IPEDS
year. It is recommended to use the \code{\link{ipeds_survey}} function instead.
}
