% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{preferences}
\alias{preferences}
\title{Function for Selecting Institution by Preferences}
\usage{
preferences(financial_aid, size, region, diversity_students, diversity_staff)
}
\arguments{
\item{financial_aid}{(numeric) Percent of undergraduate students awarded aid.}

\item{size}{(numeric) Amount of Students attending the institution. (1    = Under 1,000; 2 = 1,000 - 4,999; 3    = 5,000 - 9,999; 4 = 10,000 - 19,999; 5    = 20,000 and above; -1 = Not reported; -2    = Not applicable)}

\item{region}{(character) Region the institution resides.}

\item{diversity_students}{(numeric) Percent of non-white students.}

\item{diversity_staff}{(numeric) Percent of non-white staff.}
}
\value{
A dataframe of the institution that fit inputted preferences. For arguments that return percentages, the user can expect institutions that meet that percentage or higher.
}
\description{
Lists the names of institutions that fit inputted preferences
}
\examples{
preferences(70, 2, "New England", 20, 40)
preferences(50, 4, "Southwest", 10, 20)
}
