% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Function for importing raw data from IPEDS}
\usage{
read_data(dataset)
}
\arguments{
\item{dataset}{(character) abbreviation of dataset from IPEDS. For the list of all possible datasets to import, visit the \href{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?year=-1}{IPEDS datacenter} and refer to the \verb{Data File} column.}
}
\value{
A dataframe with desired information
}
\description{
Outputs data frame with desired information directly from IPEDS
}
\examples{
\donttest{
read_data("IC2020")
}
}
