% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recodeHighestDegreeOffered}
\alias{recodeHighestDegreeOffered}
\title{recodeHighestDegreeOffered}
\usage{
recodeHighestDegreeOffered(col)
}
\arguments{
\item{col}{the column name}
}
\description{
recodeHighestDegreeOffered
}
