% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recodeHighestLevelOfOffering}
\alias{recodeHighestLevelOfOffering}
\title{recodeHighestLevelOfOffering}
\usage{
recodeHighestLevelOfOffering(col)
}
\arguments{
\item{col}{the column name}
}
\description{
recodeHighestLevelOfOffering
}
