% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recodeLevelOfInstitution}
\alias{recodeLevelOfInstitution}
\title{recodeLevelOfInstitution}
\usage{
recodeLevelOfInstitution(col)
}
\arguments{
\item{col}{the column name}
}
\description{
recodeLevelOfInstitution
}
