% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recodeTitleIVEligibility}
\alias{recodeTitleIVEligibility}
\title{recodeTitleIVEligibility}
\usage{
recodeTitleIVEligibility(col)
}
\arguments{
\item{col}{the column name}
}
\description{
recodeTitleIVEligibility
}
