% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recodeUndergraduateOffering}
\alias{recodeUndergraduateOffering}
\title{recodeUndergraduateOffering}
\usage{
recodeUndergraduateOffering(col)
}
\arguments{
\item{col}{the column name}
}
\description{
recodeUndergraduateOffering
}
