% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relig_aff}
\alias{relig_aff}
\title{Religious Affiliations Data}
\format{
A data frame of 64 rows representing religious affiliations and 2 variables

\describe{
\item{Numeric_Code}{Numeric code representing a religious affiliation}
\item{Religious_Affiliation}{Explanation of numeric code}
}
}
\usage{
relig_aff
}
\description{
This file contains the details for religious affiliations in the Offerings 2020 (\code{offerings2020}) dataset.
}
\keyword{datasets}
