% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staff_cat}
\alias{staff_cat}
\title{Staff Categories Data}
\format{
A data frame of 65 rows representing staff categories and 2 variables

\describe{
\item{Numeric_Code}{Numeric code representing a staff category}
\item{Category}{Explanation of numeric code}
}
}
\usage{
staff_cat
}
\description{
This file contains the details for staff categories in the Administration 2020 (\code{admin2020}) dataset.
}
\keyword{datasets}
