% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{surveys}
\alias{surveys}
\title{List of available surveys to load from IPEDS}
\description{
List of available surveys to load from IPEDS
}
\references{
National Center for Educational Statistics. Integrated 
     Postsecondary Education Data System \url{https://surveys.nces.ed.gov/ipeds/}
}
\keyword{data}
