% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_exists.R
\name{url_exists}
\alias{url_exists}
\title{Returns whether the given URL is valid.}
\usage{
url_exists(x, non_2xx_return_value = FALSE, quiet = TRUE, ...)
}
\arguments{
\item{x}{a single URL}

\item{non_2xx_return_value}{what to do if the site exists but the
HTTP status code is not in the `2xx` range. Default is to return `FALSE`.}

\item{quiet}{if not `FALSE`, then every time the `non_2xx_return_value` condition
arises a warning message will be displayed. Default is `FALSE`.}

\item{...}{other params (`timeout()` would be a good one) passed directly
to `httr::HEAD()` and/or `httr::GET()`}
}
\description{
This function was posted on StackOverflow here:
https://stackoverflow.com/questions/52911812/check-if-url-exists-in-r
}
\author{
hrbrmstr
}
