% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adm2020}
\alias{adm2020}
\title{Admissions and Test Scores 2020 Data}
\format{
A data frame of 1989 rows representing hourly measurements and 39 variables

\describe{
\item{UNITID}{Unique identification number of the institution: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON1}{Secondary school GPA: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON2}{Secondary school rank: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON3}{Secondary school record: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON4}{Completion of college-preparatory program: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON5}{Recommendations: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON6}{Formal demonstration of competencies: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON7}{Admission test scores: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON8}{TOEFL (Test of English as a Foreign Language: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{ADMCON9}{Other Test (Wonderlic, WISC-III, etc.: a factor with 4 levels: \code{Required}, \code{Recommended}, \code{Neither_required_nor_recommended}, and \code{Considered_but_not_required}}
\item{APPLCN}{Applicants total}
\item{APPLCNM}{Applicants men}
\item{APPLCNW}{Applicants men}
\item{ADMSSN}{Admissions total}
\item{ADMSSNM}{Admissions men}
\item{ADMSSNW}{Admissions women}
\item{ENRLT}{Enrolled total}
\item{ENRLM}{Enrolled men}
\item{ENRLW}{Enrolled women}
\item{ENRLFT}{Enrolled full time total}
\item{ENRLFTM}{Enrolled full time men}
\item{ENRLFTW}{Enrolled full time women}
\item{ENRLPT}{Enrolled part time total}
\item{ENRLPTM}{Enrolled part time men}
\item{ENRLPTW}{Enrolled part time women}
\item{SATNUM}{Number of first-time degree/certificate-seeking students submitting SAT scores}
\item{SATPCT}{Percent of first-time degree/certificate-seeking students submitting SAT scores}
\item{ACTNUM}{Number of first-time degree/certificate-seeking students submitting ACT scores}
\item{ACTPCT}{Percent of first-time degree/certificate-seeking students submitting ACT scores}
\item{SATVR25}{SAT Evidence-Based Reading and Writing 25th percentile score}
\item{SATVR75}{SAT Evidence-Based Reading and Writing 75th percentile score}
\item{SATMT25}{SAT Math 25th percentile score}
\item{SATMT75}{SAT Math 75th percentile score}
\item{ACTCM25}{ACT Composite 25th percentile score}
\item{ACTCM75}{ACT Composite 75th percentile score}
\item{ACTEN25}{ACT English 75th percentile score}
\item{ACTEN75}{ACT English 75th percentile score}
\item{ACTMT25}{ACT Math 25th percentile score}
\item{ACTMT75}{ACT Math 75th percentile score}
}
}
\source{
Download from Integrated Postsecondary Education Data System,
\url{https://nces.ed.gov/ipeds/datacenter/data/ADM2020.zip}.
}
\usage{
adm2020
}
\description{
This file contains information about the undergraduate selection process for entering first-time, degree/certificate-seeking students. This includes information about admission considerations,  applicants, applicants that were admitted, and admitted students who enrolled. SAT and ACT test scores are included for institutions, that require test scores for admission. These data are applicable for institutions that do not have an open admissions policy for entering first-time students. Beginning with 2016, writing scores for both SAT and ACT are no longer collected. The possible values for the admission consideration variables have changed from data prior to 2016.
}
\details{
Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions.
}
\keyword{datasets}
