% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_IPEDS.R
\name{available_ipeds}
\alias{available_ipeds}
\title{Returns what IPEDS databases are available.}
\usage{
available_ipeds(dir = paste0(system.file(package = "ipeds"),
  "/data/downloaded/"))
}
\arguments{
\item{dir}{directory where IPEDS databases are to be downloaded.}
}
\value{
a data.frame with final and provisional columns indicating whether
        those databases are available for download. The downloaded column 
        indicates whether that database has already been downloaded.
}
\description{
Returns what IPEDS databases are available.
}
