% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dir_info2020}
\alias{dir_info2020}
\title{Directory Info 2020 Data}
\format{
A data frame of 6220 rows representing institutions and 47 variables

\describe{
\item{INSTITUTION_ID}{Unique identification number of the institution}
\item{INSTITUTION}{Institution (entity) name}
\item{ADDRESS}{Street address or post office box}
\item{CITY}{City location of institution}
\item{STATE}{State abbreviation}
\item{ZIP}{ZIP code}
\item{FIPS}{FIPS state code}
\item{BEA_REG}{Bureau of Economic Analysis (BEA) regions}
\item{AID_ELIG}{OPE Title IV eligibility indicator code}
\item{WEBSITE}{Institution's internet website address}
\item{ADMIN_WEB}{Admissions office web address}
\item{FINAID_WEB}{Financial aid office web address}
\item{APPLICATION}{Online application web address}
\item{NPRICE_CALC}{Net price calculator web address}
\item{VET_TUITION}{Veterans and Military Servicemembers tuition policies web address}
\item{ATH_GRAD_RATE}{Student-Right-to-Know student athlete graduation rate web address}
\item{DISABILITY_WEB}{Disability Services Web Address}
\item{INT_LEVEL}{Level of institution}
\item{INT_CONTROL}{Control of institution}
\item{UNGRAD_OFFER}{Undergraduate offering}
\item{GRAD_OFFER}{Graduate offering}
\item{HD_OFFER}{Degree-granting status}
\item{HBCU}{Historically Black College or University}
\item{HOSPITAL}{Institution has hospital}
\item{MEDICAL}{Institution grants a medical degree}
\item{TRIBAL}{Tribal college}
\item{LOCALE}{Degree of urbanization (Urban-centric locale)}
\item{OPENPUBL}{Institution open to the general public}
\item{INT_STATUS}{Status of institution}
\item{CLOSE_DATE}{Date institution closed}
\item{POST_SEC}{Postsecondary institution indicator}
\item{RPTMTH}{Reporting method for student charges, graduation rates, retention rates and student financial aid}
\item{CC_U2018}{Carnegie Classification 2018: Undergraduate Instructional Program}
\item{CC_G2018}{Carnegie Classification 2018: Graduate Instructional Program}
\item{CC18_U_PROFILE}{Carnegie Classification 2018: Undergraduate Profile}
\item{CC18_ENROLL}{Carnegie Classification 2018: Enrollment Profile}
\item{CC18_SIZE_SET}{Carnegie Classification 2018: Size and Setting}
\item{C_PROGRAMS}{Carnegie Classification 2000}
\item{INT_SIZE}{Institution size category}
\item{MULT_ORG}{Multi-institution or multi-campus organization}
\item{MULT_ORG_NAME}{Name of multi-institution or multi-campus organization}
\item{MULT_ORG_CODE}{Identification number of multi-institution or multi-campus organization}
\item{COUNTYCD}{Fips County code}
\item{COUNTYNM}{County name}
\item{CONGRESS_DIS_ID}{State and 114TH Congressional District ID}
\item{LONGITUD}{Longitude location of institution}
\item{LATITUDE}{Latitude location of institution}
}
}
\source{
Download from Integrated Postsecondary Education Data System,
\url{https://nces.ed.gov/ipeds/datacenter/data/HD2020.zip}.
}
\usage{
dir_info2020
}
\description{
This file contains directory information for every institution in the 2020-21 IPEDS universe. Includes name, address, city, state, zip code and various URL links to the institution's home page, admissions, financial aid offices and the net price calculator. Identifies institutions as currently active, institutions that participate in Title IV federal financial aid programs for which IPEDS is mandatory. It also includes variables derived from the 2020-21 Institutional Characteristics survey, such as control and level of institution, highest level and highest degree offered and Carnegie classifications.
}
\details{
Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions.
}
\keyword{datasets}
