% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIPEDSSurvey.R
\name{downloadIPEDSSurvey}
\alias{downloadIPEDSSurvey}
\title{downloadIPEDSSurvey}
\usage{
downloadIPEDSSurvey(
  surveyId,
  year,
  dir = system.file(package = "ipeds"),
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{surveyId}{the survey ID from the surveys data frame}

\item{year}{the year of the survey}

\item{dir}{the directory to download the data to. Defaults to the package directory}

\item{stringsAsFactors}{ogical: should character vectors be converted to factors? 
Note that this is overridden by as.is and colClasses, both of which allow 
finer control.}

\item{...}{other parmaters passed to \code{\link{read.csv}}}
}
\description{
Downloads a given IPEDS survey data file.
}
