\name{fallEnrollment}
\alias{fallEnrollment}
\title{Internal method.}
\usage{
  fallEnrollment(data, mappings = NULL, institutionId = NA,
    institutionName = NA, documentId = "a",
    createdDate = format(Sys.time(), "\%Y-\%m-\%dT\%H:\%M:\%S"),
    noteMessage = NA, transmissionType = "Original",
    documentType = "IPEDS")
}
\description{
  Internal method.
}

