% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIPEDSSurvey.R
\name{getIPEDSSurvey}
\alias{getIPEDSSurvey}
\title{getIPEDSSurvey}
\usage{
getIPEDSSurvey(surveyId, year, dir = system.file(package = "ipeds"),
  stringsAsFactors = FALSE, ...)
}
\arguments{
\item{surveyId}{The survey ID of the dictionary to display (i.e. surveys$SurveyID)}

\item{year}{The year of the dictionary to display.}

\item{dir}{the directory to download the data to. Defaults to the package directory}

\item{stringsAsFactors}{ogical: should character vectors be converted to factors? 
Note that this is overridden by as.is and colClasses, both of which allow 
finer control.}

\item{...}{other parmaters passed to \code{\link{read.csv}}}
}
\description{
This function will return the IPEDS survey data for the given survey and year. 
Note that the first time this function is invoked it will attempt to download 
the data file from the IPEDS website.
}
